# specs/openapi/workspace.yaml
...
openapi: 3.1.0
info:
  title: Plinth Workspace API
  version: 0.1.0
  description: |
    The Workspace service is the persistent, structured memory layer for agents. It
    provides versioned key-value storage, versioned files, immutable snapshots and
    divergent branches that can be merged back into the main timeline.

    All endpoints require a bearer token via `Authorization: Bearer <token>`. In the
    v0.1 PoC any non-empty token is accepted. Capability tokens (see
    `specs/schemas/capability-token.schema.json`) are scheduled for v0.2.

    See `CONTRACTS.md` for the prose source of truth.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:7421
    description: Local development server (default port)
...
security:
  - bearerAuth: []
paths:
...
  /v1/workspaces:
    get:
      tags: [workspaces]
      summary: List workspaces
      operationId: listWorkspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of workspaces visible to the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [workspaces]
      summary: Create a workspace
      description: |
        Creates a new workspace. Names are not required to be unique; clients
        looking for "get-or-create" semantics should list first then create.
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreate'
      responses:
        '201':
          description: Workspace created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workspaces/{ws_id}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [workspaces]
      summary: Get a workspace by id
      operationId: getWorkspace
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workspace found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [workspaces]
      summary: Delete a workspace
      description: Permanently deletes a workspace and all its contents.
      operationId: deleteWorkspace
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Workspace deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/workspaces/{ws_id}/kv:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/BranchQuery'
    get:
      tags: [kv]
      summary: List latest KV entries
      description: Returns the latest version of every key in the workspace (or branch).
      operationId: listKVEntries
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Latest entries for each key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KVEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/workspaces/{ws_id}/snapshots:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/BranchQuery'
    get:
      tags: [snapshots]
      summary: List snapshots
      operationId: listSnapshots
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All snapshots in the workspace.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [snapshots]
      summary: Create a snapshot
      description: Captures the current latest version of every key and file in the workspace (or branch).
      operationId: createSnapshot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotCreate'
      responses:
        '201':
          description: Snapshot created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token in v0.1 (any non-empty string accepted). Capability
        tokens are scheduled for v0.2.
  parameters:
    WorkspaceId:
      in: path
      name: ws_id
      required: true
      description: Workspace identifier (`ws_<ulid>`).
      schema:
        $ref: '#/components/schemas/WorkspaceIdValue'
...
    BranchQuery:
      in: query
      name: branch
      required: false
      description: |
        Target a specific branch instead of the workspace main timeline. When
        present, reads see branch-specific writes first, then fall through to the
        branch's source snapshot. Writes are scoped to the branch.
      schema:
        $ref: '#/components/schemas/BranchIdValue'
  responses:
    BadRequest:
      description: Validation or argument error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: Server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    WorkspaceIdValue:
      type: string
      description: ULID-suffixed workspace id.
      pattern: '^ws_[0-9A-HJKMNP-TV-Z]{26}$'
      example: "ws_01HZX9V5K4M7N0Q1R2S3T4V5W6"
    SnapshotIdValue:
      type: string
      description: ULID-suffixed snapshot id.
      pattern: '^snap_[0-9A-HJKMNP-TV-Z]{26}$'
      example: "snap_01HZX9V5K4M7N0Q1R2S3T4V5W6"
    BranchIdValue:
      type: string
      description: ULID-suffixed branch id.
      pattern: '^br_[0-9A-HJKMNP-TV-Z]{26}$'
      example: "br_01HZX9V5K4M7N0Q1R2S3T4V5W6"
    Workspace:
      type: object
      additionalProperties: false
      required: [id, name, created_at, updated_at, metadata]
      properties:
        id:
          $ref: '#/components/schemas/WorkspaceIdValue'
        name:
          type: string
          minLength: 1
          maxLength: 256
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          description: Free-form, JSON-serializable metadata supplied by the caller.
          additionalProperties: true
    WorkspaceCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        metadata:
          type: object
          additionalProperties: true
    WorkspaceList:
      type: object
      additionalProperties: false
      required: [workspaces]
      properties:
        workspaces:
          type: array
          items:
            $ref: '#/components/schemas/Workspace'
    KVEntry:
      type: object
      additionalProperties: false
      required: [workspace_id, key, value, version, created_at, deleted]
      properties:
        workspace_id:
          $ref: '#/components/schemas/WorkspaceIdValue'
        key:
          type: string
        value:
          description: |
            Any JSON-serializable value. `null` is allowed and distinct from a
            tombstone (which sets `deleted: true`).
        version:
          type: integer
          minimum: 1
          description: Monotonic per (workspace, key).
        created_at:
          type: string
          format: date-time
        deleted:
          type: boolean
          default: false
        branch_id:
          oneOf:
            - $ref: '#/components/schemas/BranchIdValue'
            - type: 'null'
          description: Set when the entry was written on a branch.
...
    KVEntryList:
      type: object
      additionalProperties: false
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/KVEntry'
...
    Snapshot:
      type: object
      additionalProperties: false
      required: [id, workspace_id, name, created_at, kv_versions, file_versions]
      properties:
        id:
          $ref: '#/components/schemas/SnapshotIdValue'
        workspace_id:
          $ref: '#/components/schemas/WorkspaceIdValue'
        name:
          type: string
          minLength: 1
          maxLength: 256
        message:
          oneOf:
            - type: string
            - type: 'null'
        created_at:
          type: string
          format: date-time
        kv_versions:
          type: object
          description: Map of key → captured version.
          additionalProperties:
            type: integer
            minimum: 1
        file_versions:
          type: object
          description: Map of path → captured version.
          additionalProperties:
            type: integer
            minimum: 1
        parent_snapshot_id:
          oneOf:
            - $ref: '#/components/schemas/SnapshotIdValue'
            - type: 'null'
    SnapshotCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        message:
          type: string
          maxLength: 4096
    SnapshotList:
      type: object
      additionalProperties: false
      required: [snapshots]
      properties:
        snapshots:
          type: array
          items:
            $ref: '#/components/schemas/Snapshot'
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          enum:
            - WORKSPACE_NOT_FOUND
            - KEY_NOT_FOUND
            - FILE_NOT_FOUND
            - SNAPSHOT_NOT_FOUND
            - BRANCH_NOT_FOUND
            - TOOL_NOT_FOUND
            - TOOL_INVOCATION_FAILED
            - INVALID_ARGUMENTS
            - UNAUTHORIZED
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'